Large changes
Great many changes here, restructured the repo to follow the electron-vite structure and switched to using electron-vite as the build pipeline. This simplifies things and has the benifit of being easy to breakpoint debug via VS code. TODO remove the references to outdated build pipelines such as webpack and babel when I have time.
New functionality as well, switched to using a database selector and a credentials logon screen with support for saving (encrypted) credentials.
Also added the ability to stack multiple diff frames at once, allowing them to be handily packaged up for export.